FIx bug where can't startup due no pem files to be removed

+++ frankenphp/docker-entrypoint.sh
@@ -61,8 +61,11 @@ if [ "$1" = 'frankenphp' ] || [ "$1" = 'php' ] || [ "$1" = 'bin/console' ]; then
 
 		if [ -z $JWT_PASSPHRASE ] || [ -z "$(find ./config/jwt -iname 'public.pem' -print -quit)" ]; then
 			if [ -z $JWT_PASSPHRASE]; then
-				rm ./config/jwt/public.pem
-				rm ./config/jwt/private.pem
+
+				if [ "$(find ./config/jwt -iname 'public.pem' -print -quit)" ]; then
+					rm ./config/jwt/public.pem
+					rm ./config/jwt/private.pem
+				fi
 
 				echo "Generating Passphrase"
